Move Pointer/Ref next to StreamedObject
