+++ src/rootfilespec/bootstrap/streamedobject.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass
 from enum import IntEnum
-from typing import Optional, TypeVar
+from typing import Generic, Optional, TypeVar, get_args, get_origin
 
 from rootfilespec.dispatch import DICTIONARY, normalize
 from rootfilespec.serializable import serializable
-from rootfilespec.structutil import Members, ReadBuffer, ROOTSerializable
+from rootfilespec.structutil import Members, MemberSerDe, ReadBuffer, ROOTSerializable
 
 
 class _StreamConstants(IntEnum):
@@ -76,7 +76,7 @@ class StreamHeader(ROOTSerializable):
         return cls(fByteCount, fVersion, fClassName, fClassRef), buffer
 
 
-def read_streamed_item(buffer: ReadBuffer) -> tuple[ROOTSerializable, ReadBuffer]:
+def read_streamed_item(buffer: ReadBuffer):
     # Read ahead the stream header to determine the type of the object
     itemheader, _ = StreamHeader.read(buffer)
     if itemheader.fByteCount == 0 and itemheader.fClassRef is not None:
@@ -178,3 +178,57 @@ class StreamedObject(ROOTSerializable):
     def read(cls: type[T], buffer: ReadBuffer) -> tuple[T, ReadBuffer]:
         members, buffer = _read_all_members(cls, buffer)
         return cls(**members), buffer
+
+
+class Ref(Generic[T]):
+    """A class to hold a reference to an object.
+
+    We cannot use a dataclass here because its repr might end up
+    being cyclic and cause a stack overflow.
+    """
+
+    obj: Optional[T]
+    """The object that is referenced."""
+
+    def __init__(self, obj: Optional[T]):
+        self.obj = obj
+
+    def __repr__(self):
+        label = type(self.obj).__name__ if self.obj else "None"
+        return f"Ref({label})"
+
+    @classmethod
+    def read_as(cls, ftype: type[T], buffer: ReadBuffer):  # noqa: ARG003
+        (addr,), _ = buffer.unpack(">i")
+        if not addr:
+            buffer = buffer[4:]
+            return cls(None), buffer
+        if addr & 0x40000000:
+            # this isn't actually an address but an object
+            addr &= ~0x40000000
+            buffer = buffer[addr + 4 :]
+            return cls(None), buffer
+            # obj, buffer = ftype.read(buffer)
+            # return cls(obj), buffer
+        # TODO: finish Pointer implementation
+        return cls(None), buffer[4:]
+
+
+class Pointer(MemberSerDe):
+    def build_reader(self, fname: str, ftype: type):
+        if (origin := get_origin(ftype)) is not Ref:
+            msg = f"Pointer() only can be used with Ref, got {origin}"
+            raise ValueError(msg)
+        (ftype,) = get_args(ftype)
+        if not issubclass(ftype, StreamedObject):
+            msg = f"Pointer() only can be used with Ref[StreamedObject], got {ftype}"
+            raise ValueError(msg)
+
+        def read(members: Members, buffer: ReadBuffer) -> tuple[Members, ReadBuffer]:
+            obj, buffer = Ref.read_as(ftype, buffer)
+            members[fname] = obj
+            return members, buffer
+
+        return read
+        # msg = f"Pointer to address {addr} not implemented"
+        # raise NotImplementedError(msg)
