+++ src/rootfilespec/dynamic.py
@@ -5,6 +5,7 @@ from rootfilespec.bootstrap.TStreamerInfo import (
 )
 from rootfilespec.dispatch import DICTIONARY, normalize
 
+# TODO: make a generated/__init__.py file out of this
 PREAMBLE = """# Generated by streamerinfo_to_classes
 
 from typing import Annotated, Optional
@@ -16,8 +17,6 @@ from rootfilespec.bootstrap.TKey import DICTIONARY
 from rootfilespec.container import (
     FixedSizeArray,
     BasicArray,
-    Pointer,
-    Ref,
     StdVector,
 )
 from rootfilespec.structutil import Fmt
